pricing_data/requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
defusedxml==0.7.1
coreapi==2.3.3
coreschema==0.0.4
django-filter==23.2
drf-yasg==1.21.5
Django==4.1.7
MarkupSafe==2.1.2
django-extensions==3.2.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
inflection==0.5.1
h11==0.14.0
httpie==3.2.1
itypes==1.2.0
idna==3.4
Jinja2==3.1.2
packaging==23.1
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.2
outcome==1.2.0
pandas==2.0.0
psycopg2==2.9.6
Pygments==2.14.0
pyscopg2==66.0.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.2
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
undetected-chromedriver==3.4.6
urllib3==1.26.15
websockets==11.0.2
wsproto==1.2.0
gunicorn==20.1.0
uritemplate==4.1.1